+++ packages/theme/src/prebuilt/ts/color-tokens.ts
@@ -4,6 +4,16 @@
  */
 
 export default {
+	transparent: [
+		'background-interactive-brand-weak',
+		'background-interactive-brand-weak-disabled',
+		'background-interactive-error',
+		'background-interactive-error-disabled',
+		'background-interactive-error-weak',
+		'background-interactive-error-weak-disabled',
+		'background-interactive-neutral-weak',
+		'background-interactive-neutral-weak-disabled',
+	],
 	'primary-bgFill1': [ 'background-interactive-brand-strong' ],
 	'primary-fgFill': [
 		'foreground-interactive-brand-strong',
@@ -71,7 +81,11 @@ export default {
 	'error-stroke4': [ 'stroke-interactive-error-active' ],
 	'error-stroke1': [ 'stroke-surface-error' ],
 	'bg-surface2': [ 'background-surface-neutral' ],
-	'bg-surface5': [ 'background-interactive-neutral-strong-disabled' ],
+	'bg-surface5': [
+		'background-interactive-brand-strong-disabled',
+		'background-interactive-error-strong-disabled',
+		'background-interactive-neutral-strong-disabled',
+	],
 	'bg-surface4': [ 'background-interactive-neutral-weak-active' ],
 	'bg-surface3': [ 'background-surface-neutral-strong' ],
 	'bg-fgSurface4': [
@@ -84,6 +98,10 @@ export default {
 		'foreground-interactive-neutral-weak',
 	],
 	'bg-fgSurface2': [
+		'foreground-interactive-brand-disabled',
+		'foreground-interactive-brand-strong-disabled',
+		'foreground-interactive-error-disabled',
+		'foreground-interactive-error-strong-disabled',
 		'foreground-interactive-neutral-disabled',
 		'foreground-interactive-neutral-strong-disabled',
 		'foreground-interactive-neutral-weak-disabled',
@@ -101,6 +119,8 @@ export default {
 	'bg-stroke2': [
 		'background-thumb-neutral-disabled',
 		'background-track-neutral',
+		'stroke-interactive-brand-disabled',
+		'stroke-interactive-error-disabled',
 		'stroke-interactive-neutral-disabled',
 		'stroke-surface-neutral',
 	],